pyproject.toml
[project]
name = "news-trading-bot"
version = "0.1.0"
description = "News-driven trading bot for US equities (paper trading)"
requires-python = ">=3.11"
dependencies = [
    # Broker
    "alpaca-py>=0.28.0",
    # News
    "finnhub-python>=2.4.0",
    "feedparser>=6.0.0",
    "aiohttp>=3.9.0",
    # NLP / Sentiment
    "transformers>=4.40.0",
    "torch>=2.2.0",
    # Market data & indicators
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    # Storage
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    # Config
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "pyyaml>=6.0.0",
    # Alerts
    "discord-webhook>=1.3.0",
    # Utils
    "structlog>=24.1.0",
    "cachetools>=5.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.3.0",
]

[project.scripts]
trading-bot = "src.main:main"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
